requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-language==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
libcst==0.3.16
mypy-extensions==0.4.3
praw==7.1.0
prawcore==1.5.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
selenium==3.141.0
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
